backend/requirements.txt
Django==5.2
psycopg2-binary==2.9.10
graphene-django==3.2.3
django-graphql-jwt==0.4.0
django-cors-headers==4.7.0
python-dotenv==1.1.0
requests==2.32.3
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
